rename files and fix sonar warning
